Merged: [turboshaft] Avoid putting FrameState constant bitcast input in register
When constants should be used as Smi as FrameState input, we used to
convert Int32Constant to tagged by using a BitcastWordToTagged, but a
recent CL (https://crrev.com/c/5083021) made that more precise,
generating a BitcastWordToTaggedSigned instead.

When the Instruction Selector visits a BitcastWordToTagged, it emits a
Nop with a DefineSameAsFirst (which seems to be some kind of
GapMove). However, when it visits BitcastWordToTaggedSmi, it doesn't
emit anything and just does a renaming.

For a reason that's not clear to me, the register allocator does not
want to spill registers holding constants that come from renamings,
but is happy to spill registers holding constants coming from
GapMoves.

This lead to the register allocator running out of registers when
allocating registers for FrameState inputs, because it was using one
register per integer constant input and didn't want to spill any of
them.

This CL fixes the issue by adding a special case when generating frame
states in the Instruction Selector: BitcastWordToTaggedSigned with
Word32 constant inputs now lead to using a Constant rather than using
the BitcastWordToTaggedSigned itself.

Bug: v8:12783, chromium:1520697
(cherry picked from commit a5e22070afecdb18db45872ef0eb93947085576d)

Change-Id: I5da8a10bc9ffd95511f2237488178fe8066ba65a