[maglev] Do not canonicalize smi from holey double by default
We read a holey double only if the feedback has seen a hole.
Most likely we do not need to convert the result to a smi.

Since smi canonicalization is actually expensive, we avoid it
by default. If we need to read the value as a Smi, we force
canonicalization.

Phis will still try to canonicalize.

Bug: v8:7700, v8:14568
Change-Id: I3bffdfb9e62b2a24eab988d1b704f6f1a4eb06fc