[wasm] Introduce WasmDispatchTable
This CL introduces the WasmDispatchTable alongside the existing
WasmIndirectFunctionTable.
The WasmDispatchTable lives in trusted space, so it improves security by
making it much harder to manipulate signatures or call targets. It also
improves performance because the stored information is close together
per index, requires less code for loading and touching fewer cache
lines.

For now the "ref" stored in the WasmDispatchTable is still either a
WasmApiFunctionRef or a WasmInstanceObject. After the refactoring is
complete, the WasmApiFunctionRef will also be stored in trusted space
and we will store either the WasmApiFunctionRef or the
WasmTrustedInstanceData directly. This will make the code for indirect
calls even smaller.

Future CLs will first start to fill the dispatch table alongside the
existing indirect function tables, then switch over all compilers to use
the dispatch table, and then finally remove the indirect function
tables.

R=ahaas@chromium.org
CC=saelo@chromium.org

Bug: v8:14564
Change-Id: Ib965a653ac391ba87117ccf9e913c173fcea2866