[turboshaft][wasm] WasmGCTypeReducer: Fix bottom type in snapshot merge
This is a very similar change to https://crrev.com/c/5204427 .
Back then the assumption was that bottom types can not happen in the
snapshot merging for blocks marked as unraechable.
As part of the analysis of https://crbug.com/1514072 it was also
uncovered that reachability tracking doesn't propagate causing
predecessors to be marked reachable which in fact are unreachable.

This CL fixes the handling of bottom types. Once reachability fully
propagates, bottom types shouldn't appear for reachable blocks in the
snapshot merging. Still, for defensive reasons it might still be
valuable to keep these checks alive.

Fixed: chromium:1521371
Bug: v8:14108
Change-Id: I720fac5d2cd1e3f2b6933e9797d2881385b52f96