[sandbox] First step towards fine-grained entrypoint tags for builtins
Many builtins are only used internally (e.g. called directly from a
compiled function or another builtin) and do not expect to be called
directly from script (e.g. through a JSFunction). Until now, however,
they could still be invoked from within sandbox as they were exposed
through the code pointer table (CPT).

To avoid this, CodeEntrypointTags should be used: these are XORed into
the top bits of the entypoint pointer in the CPT, and the caller must
untag the entrypoint pointer with the expected tag before jumping to it.
On Arm64, instead of using XOR, the entrypoint pointers could, in the
future, also be signed using PAC, which may be more efficient.

This CL now implements the basic machinery for assigning fine-grained
code entrypoint tags to (groups of) builtins. It then assings a new
entrypoint tag to the IC handler builtins. Other builtins will be
handled in subsequent CLs.

Bug: chromium:1507800
Change-Id: I6bfa22d541905219f50f4a6197432d25f056a229