[heap] Fix cppgc architecture macros
Most code was using V8_TARGET_* when really this macro is used for
determining the JIT target. All of Oilpan runs on the host though, so
we should use the V8_HOST_* macros.

A side effect of this is fixing a cross-compilation issue. Previously,
when setting target_cpu=x64 on arm64, we would get write barriers in
cppgc's public API wrong because the macro override is not propagated
to the API and the information about guard pages got out of sync.
Using the host macro fixes this because it is never overriden using GN
(but only via compilers directly).

Bug: chromium:1520621
Change-Id: Iba034f53b90b6f76f24d57f584e957cdde9d2177