[maglev] Inline the TypedArray.prototype.length accessor
TypedArray.prototype.length returns the integer length of the
TypedArray. We can inline this getter into a simple field load for
non-resizable typed arrays.

This unfortunately requires expanding the IntPtr representation to be a
"first-class" representation in the graph, with all the conversions,
since it can now appear as an arbitrary value in the graph and not only
an input into bounds checks. To keep things explicit about the value range
of this representation (0 ≤ value ≤ max_safe_int), rename it to
TypedArrayLength.

Drive-by fix BuildLoadTypedArrayLength to bail out fully on RAB/GSAB,
since even the uncached LoadTypedArrayLength is still unsafe.

Bug: v8:7700
Change-Id: Iad010d65f1c1ac41627abf45fe7a4f3838e34581