[riscv][heap] Check that we do not allocate in fast c calls
Port commit 391a474d7b829ea1a44d1c25dceeded421a3e023

Fast c calls are not allowed to allocate or start a GC. Add DCHECKs
and CHECKs to ensure that this is indeed the case through all of V8.

Wasm-c-api was using fast c calls with allocation as well. We could
either disable the checks for it or simply do not set
fast_c_call_caller_fp for those calls. I chose the latter approach
because then we can ensure the (D)CHECKs hold all the time.

Bug: chromium:1523607


Change-Id: I2fc6e21a8f438e61b121cba6cb03669c89c91920