[turboshaft] Avoid +0 addition in LoadStoreSimplification
We were occasionally introducing a "+ 0" addition in
LoadStoreSimplificationReducer. In most phases, this isn't an issue
because MachineOptimizationReducer will probably run afterwards and
get rid of such useless additions, but LoadStoreSimplification runs in
the last phase, which doesn't have a MachineOptimizationReducer.

We thus ended up emitting additional machine instructions in such
cases, which lead to a few performance regressions.

Bug: v8:12783
Change-Id: I8254d7ff29df801d4530b1441af77ac7f34c4bd5
Fixed: chromium:1524335