[sandbox] Allow double-initialization of EPT entries
Previously, we tried to avoid double-initialization of EPT entries as it
would lead to slightly inefficient table compaction: we might relocate
an entry that doesn't need to be relocated. However, we can avoid this
simply by checking whether the current entry is already below the
evacuation area (where it must be if it has been newly allocated). What
remains is the inefficieny of (potentially) allocating a new table
entry, although depending on the context that can be hard to avoid if
the entry is cleared in between.

The problem with the previous DCHECKs was that they were unactionable:
they would only tell us that a slot was re-initialized somewhere, but
couldn't point us in any direction for where to look. As such, with this
CL we now simply allow double-initialization and adjust the algorithm to
deal with that efficiently.

What does remain though is the limitation that we must not share EPT
entries between multiple objects. So far this has never been an issue
though (e.g. as there are no APIs that would do this).

Bug: chromium:41482529
Change-Id: I96185130524ffc2a9bfcd90c709f1fe40719f0a3