Merged: [wasm] Add generic wasm-to-js wrapper for invalid signatures
So far a specific wasm-to-js wrapper got compiled when a wrapper was
requested for a signature that was not compatible with JS, e.g. a
signature that contains a simd. This CL introduces a generic wrapper for
invalid signatures that throws an exception unconditionally.

There was a code path where the generic wrapper was used for invalid
signatures. This code path now also  uses the new wrapper for invalid
signatures.

R=thibaudm@chromium.org

Bug: 323694592
(cherry picked from commit eacf7f40e7dd2ca11b6e2f63dd8e8bd681ca91c4)

Change-Id: Ia922bd2cc0a377b829bbfd6a361c63e1e6f03d7a