Merged: [wasm] Add generic wasm-to-js wrapper for invalid signatures
This merge contains two CLs:

CL1:
[wasm] Add generic wasm-to-js wrapper for invalid signatures

So far a specific wasm-to-js wrapper got compiled when a wrapper was
requested for a signature that was not compatible with JS, e.g. a
signature that contains a simd. This CL introduces a generic wrapper for
invalid signatures that throws an exception unconditionally.

There was a code path where the generic wrapper was used for invalid
signatures. This code path now also  uses the new wrapper for invalid
signatures.

CL2:
[wasm] Fix no-generic-wrapper case in GetOrCreateWasmInternalFunction

The existing code always used the generic wrapper, even when the
generic wrapper was disabled. With this CL we use the generic wrapper
only if it is enabled.

R=thibaudm@chromium.org

Bug: chromium:323694592
(cherry picked from commit eacf7f40e7dd2ca11b6e2f63dd8e8bd681ca91c4)

Change-Id: If10bf51fc5fd05c68b6c89affefcbe7d950b4ad7