[compiler] Always emit safepoint for fast C calls
This CL starts recording safepoints for fast c call sites in order
to eventually allow allocations in them. The safepoint needs to be
recorded exactly after the call site such that the StackFrameIterator
is able to find it. For this MacroAssembler::CallCFunction now
returns the pc_offset for the actual raw call instruction. This
return value is then used as input to RecordSafepoint().

Wasm-C-API was using an approximate return address of that call site,
so far this was okay because the safepoint was recorded there as well.
In order to not emit two different safepoints for each fast c call,
MacroAssembler::CallCFunction also optionally takes a label as
argument. This label is bound to the exact return address of the
raw call instruction.

Bug: 1523607
Change-Id: I73c458b590e2dad938f505308066bfe3c07e9e6e