Refactor to get rid of GetPromiseOnStackOnThrow()
Changes catch prediction to work more like async stack traces without
relying on the promise stack. This will allow the promise stack to be
deleted in a future cl.

In the course of this I have also fixed
handling of exceptions in the Promise constructor, cases
where promise rejections were thought to be handled by synchronous
try/catch blocks, cases where promise rejections are forwarded to
promises that are not pending (e.g. Promise.race that is already
fulfilled.), and cases where a promise has multiple listeners,
some of which handle it and some of which leave it uncaught.

Bug: b:40283985, b:40283987
Change-Id: If2c62af8c183f27a89bd4ca8aa95b3112a6a59cd