[sandbox] Use Smi::zero() as cleared position table value
The source_position_table field of a BytecodeArray can now only have two
legitimate values: a ByteArray (follow-up CLs will turn this into a
TrustedByteArray) and Smi::zero(), indicating that the field is
unavailable/uninitialized. With this, it becomes impossible to determine
whether the position table is missing because it was never computed, or
because there was an error during the computation. However, this does
not currently appear to be relevant.

This change prepares for a follow-up change that turns the position
table ByteArray into a TrustedByteArray, and the references to it into
protected pointers. For that, all possible values must either be Smis or
trusted objects (located in trusted space).

Bug: chormium:40927183
Change-Id: Idd93a88223be6e27fb393c6c4eea3fb00bab04f0