[sandbox] Mark trusted roots as immovable
This will prevent the garbage collector from moving them during
compacting GC, which in turn allows these roots to be accessed when
handle dereference would otherwise be forbidden. This is necessary for
follow-up CLs that replace empty_byte_array (a read-only root) with
empty_trusted_byte_array (a trusted root).

Bug: b/40069826
Change-Id: I04ff9888cd3a8c92e6f2f0f68c2e0fbe4512aa53