[wasm-imported-strings] Fix IsArrayNewSegment in unreachable code
The node we query is invalid in unreachable code, so we must check
for that before fetching it from the graph.

Bonus: running the regression test in Turbofan revealed an unrelated
issue having to do with reachability of catch blocks after replacing
a well-known import, so this CL includes a fix for that too.

Fixed: chromium:324690505
Change-Id: If648faf7dd466ed857d723eec44fc01fc7a6a6fc