[turboshaft] Remove END_IF macro and if stack
Previously, the IF/ELSE_IF/ELSE macros were implemented as separate
single-iteration for loops, which communicate via pushing/popping off an
if stack, and require an explicit END_IF to terminate them (since we
don't know ahead of time if an IF needs and ELSE).

Now, they are implemented as an if-else in a single two-iteration for
loop, which wraps both IF and ELSE, allowing the for loop to maintain
state and handle teardown, removing the need for an if stack or an
END_IF macro. Notably, ELSE_IF becomes a separate ELSE and IF (similar
to C++ semantics), with the inner if nested inside the outer else.

Additionally, clean up clang formatting of ELSE to not introduce a new
line.

Change-Id: I41705c5550b6fdcae6150d89a2c499bd855c743f