[fuzzer] Support arbitrary number of arguments
The standalone binaries generated for libfuzzer-based fuzzers currently
require to pass exactly one argument.
For fuzzers that can also be invoked similarly to d8, this breaks the
"bad build" check in Clusterfuzz, where the binary is being invoked with
no arguments.

Also, if unknown flags are passed, we try to open a file with that name
and fail, which is also interpreted as a bad build.

This CL fixes both issues by allowing an arbitrary number of arguments,
including 0, and ignoring additional arguments that start with '--'
(after printing a warning).

R=machenbach@chromium.org

Bug: 325533621
Change-Id: I2d2eab1ba5fc67b7fda19e262201a2b859fef590