[turboshaft][wasm] Fix load type for LOAD_TAGGED_ROOT
The root pointers are stored as uncompressed pointers, therefore
loading from the roots table should load with pointer size.
This breaks on big endian platforms otherwise as we are loading the
wrong half of the pointer then.

This change also simplifies the existing macros to load a root value
with a single macro that uses the root index information to figure out
which load to generate and whether a bitcast is required.

Change-Id: I4b419a7707bf811691dfeebe1e6e89f225dd7870