[sandbox] Prevent OOB access into elements_accessors_ array
The global elements_accessors_ array is accessed with an ElementsKind
read from within the sandbox (from a Map, typically) as index. As such,
an attacker can fully control the index. Previously, the array had only
as many entries as there are element kinds, which meant that an attacker
could access the array out-of-bounds. This CL now fixes that by making
the array 256 elements large. As the ElementsKind is a uint8_t under the
hood, and because the array contains pointers, this now guarantees that
any out-of-bounds access (index > kElementsKindCount) leads to a
(harmless) nullptr dereference crash.

Bug: b/326109685
Change-Id: I6cc8edec49b9e7bb280ddd04d42341967898ca30