[heap] Reorder GC selection conditions
When selecting a garbage collector for a GC cycle, we would force a full
GC if incremental marking was complete and allocation limit is overshot,
even if a minor GC was requested.
This should never happen with MinorMS due to separate GC phases.
Reorder the conditions so that allocation limit overshot never needs to
be checked for MinorMS.

Drive-by: Remove unused Heap::IsMajorMarkingComplete method.

Change-Id: I478c013380abc1769e39553bc6ec083acec0fb7a