[flags] Example fuzz test
This adds a fuzz test demonstrating the use of the InRegexp domain
and applying this to stress flag setting and implication processing.

This test is more of a fuzz-test example than it is an actual
valuable fuzz test from which we expect any errors to be found.

Stress tested locally to ensure it's not running into anything
obvious.

Bug: 41497291, 40941204
Change-Id: I197b1555e09bf01173f241e576d48b2da3e29546