[wasm-gc] Relax static type reqs for call_indirect
As discovered on https://github.com/WebAssembly/binaryen/pull/6336,
we erroneously required that the immediate type of a call_indirect
instruction be a subtype of the referenced table's type, but there
is no such requirement in the spec. This patch drops the check.
This change is backwards-compatible because it makes V8's behavior
strictly more permissive.

Bug: v8:9495
Change-Id: I198822e3d1ef8d8dd349fa92ed9e49d043d5d192