Allocate less times for heap worklists
After observing that several JS contexts in independent threads
belonging to the same isolate, would make Isolate::MeasureMemory
very slow, tracing showed that this was because of many work lists
being allocated through partition alloc in Chromium, so spinning
locks would make that slow.

To avoid that problem, reduce the number of allocations by storing
worklists in an std::vector preallocated with reserve(), and index
it using the V8 flatmap implementation in hashmap.h.

Change-Id: I9d553e906c1a63b53f41e37a49dc6e7257964abf