[wasm] Skip random module generation code in official builds
The code now lives in src/ so the GenerateRandomWasmModule runtime
function can use it; this increases the size of libv8.so though, so skip
this code in official builds. To avoid compile failures we still define
the GenerateRandomWasmModule function, but make it crash with an
"UNIMPLEMENTED" message if it is ever invoked.
The runtime function will just not be defined in official builds.

R=jkummerow@chromium.org

Bug: v8:14637
Change-Id: I306e73b74152f7432180a407e08d1cd381bb2372
Cq-Include-Trybots: luci.v8.try:v8_linux64_official_rel