[maps] Treat prototype transitions like elements transitions
Say we have the following code:

    var x = new foo()
    x.a = 1.1
    x.__proto__ = bar.prototype


Currently we produce the following transition tree, where double
arrows are strong and single arrows are weak pointers:

    foo {} <==--> foo {a} ----> bar {a} <==== x
                  ~~~~~~~

Assuming x to be a live object, then the underlined subtree is still
only weakly reachable from x. In other words it can be collected and
if we later create a second such object, the existing `bar {a}` map is
not reachable anymore. We re-create the whole subtree and the new
object will have a fresh (but identical) map.

The problems are:
1. Accidental polymorphism.
2. Unnecessarily duplicated maps.
3. Prototype transitions must generalize all fields
   (because we have no back-pointers and thus field updates cannot be
    properly back-propagated).
4. Instances must be migrated on prototype transitions.

The solution here is to have a strong back-pointer for prototype
transitions

    foo {} <==--> foo {a} <==--> bar {a} <==== x

This fixes issues (1)-(4) for the cost of potentially leaking
`foo.prototype`.


Change-Id: Ibc26775a3d0718cfac3354b9ca3409d29a3a5ad1