[sandbox] Validate SharedFunctionInfo::builtin_id
The builtin id is read from the heap so must be considered untrusted
when accessed. However, other code, such as SharedFunctionInfo::GetCode,
treats the builtin id as a trusted value (e.g. for indexing into the
Isolate's builtin_table). To support this, we now simply validate the
builtin id after reading it from the heap.

Bug: b/327732554
Change-Id: I19fe66d41deb51fb782d954f546bc0cc74f157af