[compile hints] Don't delete compile hints when they're fetched
We already [try to] produce crowdsourced and local compile hints at the same time,
so we cannot clear them when we fetch them.

In addition, if we want to produce local compile hints early and augment them later,
we'd need to keep the earlier hints around.

The extra memory consumption of keeping the data around longer should be negligible in
comparison to compiled code.

Bug: b/40286622
Change-Id: I542a7ab99f502ea36f3a4b42f9d7c9c1761e81f9