[turboshaft][test] Add codegen tester (disabled)
This CL just copy-pastes Turbofan codegen-tester and
test-branch-combine without any changes beyond disabling them with
`#if 0`.

This looks like a fairly useless commit, but now we can port the
codegen-tester and test-branch-combine to Turboshaft, and reviewing
will be much easier (because we'll see the delta from Tubofan to Turboshaft, which we wouldn't see easily if we were to port these
files in the same CL that creates them).

Bug: v8:12783
Change-Id: Ie761e46223de7a0c0f4d171e603b4268f1346d57