[wasm-compile-fuzzer] Modularise the fuzzer PART 1
This change templatizes the GenerateRandomWasmModule() function and the
WasmGenerator class with the WasmModuleGenerationOptions struct.

The WasmModuleGenerationOptions will be used in the upcoming changes to
specify which subset of expressions should be generated (e.g. SIMD,
wasmGC..).

This change should not change any behaviour, it only prepares the code
to modularise the fuzzer in upcoming changes

Change-Id: I2d5922ea4437bfe865518f17c41d8a1611f515f2