[multimodal] Switch to concurrent optimizations for efficiency mode
Real-world data seems to suggest that for very constrained devices
synchronous compilation on background tabs actually increases main
thread congestion. An explanation would be that very long compilation
keeps on blocking when we switch a tab into the foreground. Or, that
in general background tabs do make less progress when blocked by
compilation. For a new round of experiments we thus want to switch back
to concurrent compilation.

Bug: chromium:41485666
Change-Id: I8f6790ba760e77c85ce180ae78574548cb33c9ba