Fix handling of cleared function_data in SFI::HasBuiltinId()
When calling SharedFunctionInfo::HasBuiltinId() from a background
thread, it can happen that a SetData() call on the foreground thread
transitions that SFI from function_data to trusted_function_data. In
that case, HasBuiltinId() may see a Smi value for function_data even
though the SFI is not a builtin SFI. With this CL, this scenario is now
handled correctly and HasBuiltinId() will return false.

Bug: chromium:328294186
Change-Id: I8198002030cb9e6316091580c00881bff3043906