[isolate-groups] Beginnings of isolate groups
Up to now, for users that want pointer compression on >4GB heaps, the
official answer has been to reserve one pointer cage per isolate.
However this is perhaps wasteful of virtual memory space; it would be
useful to offer an intermediate point, where the user can explicitly
group isolates together, to allow them to be allocated in the same
ranges of virtual memory.

This patch takes an initial step in that direction by refactoring the
internal IsolateAllocator into IsolateGroup.  The next patch will add
API and tests.

Bug: v8:14625

Change-Id: I75c9a120df467b511a0f1a5500caaf28d41223c0