[arm64] Fix veneers for large code objects
Consider the following scenario:
  [previous code]
  b.eq label     # instruction 0
  [20KB of code]
  tbz label      # instruction 1
  [20KB of code]
  tbz label      # instruction 2
  [12KB of code]
  <<< position 1
  [around 960KB of code]
  <<< position 2
  [more code]
  label:
When we reach "position 1", instruction #1 is reaching the end of its
encodable jump distance, so we emit a veneer for it. We would like to
rewire the instruction chain so that instruction #2 points at #0, but
that is also out of encodable range, so we make instruction #0 branch
to the veneer as well, and mark instruction #2 as the end of the chain
of linked branch instructions referring to the label.
However, instruction #0 is still in the list of unresolved branches,
from which we must remove it; otherwise when we reach position 2,
where instruction #0's encodable jump distance ends, we will try to
emit a veneer for it, which involves trying to find it by walking the
label's chain of linked instructions -- where we won't find it, because
that chain now ends at instruction #2.

This bug was flushed out by --turboshaft-wasm, but is neither Wasm nor
Turboshaft specific (e.g. in a local repro that is too slow to check
in as a regression test, I can trigger it with Turbofan as well).

Fixed: 328738218
Change-Id: Ic98eaddbecf52bce72e25ea256e1f00d33a29b66