[fuzztest] Add json fuzzer
This adds an exemplary fuzz test for json, inspired by the existing
libfuzzer json fuzzer.

The inputs are structurally aware and focus on valid json inputs,
generated with the help of jsoncpp (while the existing libfuzzer
implementation would work on arbitrary strings mostly leading to
parse errors). In the future, we could also add another fuzz test
with arbitrary inputs.

Domains for FuzzTest are constructed with utilities from:
https://github.com/google/fuzztest/blob/main/doc/domains-reference.md

Bug: 40941204
Change-Id: I76c7337ffb1a5a530dedcfba092e0d9260cdec66