[maglev] Analyze allocated object dependencies until a fixpoint
... at the end of graph building.

This fixes the issue when we have 3 allocations (A, B, C)
and the dependencies (A, C) and (B, C).

Assuming A and C are not escaping, but B is.

If we process A first, we will remove the node, since it
is non-escaping.

But later, we process B, which forces C to escape.
Therefore, A needs to escape as well!

Bug: 329474361, 326091155
Change-Id: I4682c0b78f9514a2aef804d55a81ba9954209f79