[maglev] Use a DisjointSet for allocation dependencies
... and analyse them after AnyUseMarkingProcessor, which will
delete most of unused nodes.

The non escaped allocations and its stores are them removed from the
graph in DeadNodeSweepingProcessor.

If the store makes nodes unused, it is then too late to remove them.

This fixes the issue when we have 3 allocations (A, B, C)
and the dependencies (A, C) and (B, C).

Assuming A and C are not escaping, but B is.

If we process A first, we will remove the node, since it
is non-escaping.

But later, we process B, which forces C to escape.
Therefore, A needs to escape as well!

This CL also allows multiple dependencies.

This CL also simplifies the HasEscaped predicate:
After EscapeDependentAllocationsIfNeeded, we use a simple boolean
flag to denote escaped objects.

Bug: 329474361, 326091155
Change-Id: I4682c0b78f9514a2aef804d55a81ba9954209f79