[wasm] Rename protected fields on WasmTrustedInstanceData
This CL introduces the convention that protected fields start with a
"Protected" prefix to make it easier to know in compilers how to load
them.

Liftoff and Turboshaft already have macros for loading protected fields
from the instance, those automatically add that prefix now. For Turbofan
we introduce a similar macro in this CL.

We also slightly extend the Turboshaft macro to include the V::Cast.

R=ahaas@chromium.org

Bug: v8:14564
Change-Id: Id3811644fb8f72f0a64026de58c964e1e248605c