[wasm-compile-fuzzer] Change order of types in GenerateInitExpr
This CL changes the constant order of some types that we choose
randomly from. This is becuase if there are no more random bits
available, the default choice is element 0 which is a
HeapType::kStruct. To generate the first struct we could have a deep
stack of recursive calls GenerateInitExpr() and
GenerateStructNewInitExpr.
After this change the type 0 is I32 that can be easily initialised.

Change-Id: Id1550f7237f4d9d854928625de986a2b1777265e