[test][asmjs] Avoid data race in test by increasing sampling interval
There seems to be a data race in the CPU profiler when two samples get
collected at the same time. This happens in tests where samples get
collected explicitly through a test API, in addition to the samples that
get collected periodically by the profiler.

With this CL we increase the sampling interval of the profiler, so that
the test finishes before the profiler wants to collect its sample.

This CL also reverts a fix that was attempted before but does not, as
far as I understand help in all cases.

R=jkummerow@chromium.org

Bug: v8:14549
Change-Id: I9608776065f30f7f98040f418235e1ca60f8deae