[wasm] Reuse WasmDispatchTable for imports
The ref and target for imports is currently stored in two different
trusted data structures (imported_function_refs and
imported_function_targets).
Storing them in the same data structures avoids one load per direct call
to an import and improves cache locality.

For simplicity we re-use the WasmDispatchTable for this, even though we
do not use the signature field. This wasted four bytes per entry. If we
want to gain this back, we would need to introduce a separate class
which only stores refs and targets.

R=ahaas@chromium.org

Bug: v8:14564, chromium:330766602
Change-Id: Id8e8f8fb9a659e892341a5b250262e9bab2df05d