[heap, sandbox, api] Introduce CppHeapPointer and friends
This introduces a CppHeapPointer that's always a EPT handle on builds
using pointer compression and a full pointer otherwise.

This means that even in builds with pointer compression and without
sandbox we are using an EPT to access this external pointer. This is
necessary to allow concurrent marking for such pointers as they would
otherwise be unaligned 64-bit pointers.

Bug: chromium:328117814
Change-Id: I08d9c2463d7b7703a21169842e241c46d0375138