[sandbox] Fix handling of invalid enum value in ElementsKindToString
When an invalid enum value is passed into ElementsKindToString, we
previously had undefined behavior as the function would not return a
value at the end. This could lead to a sandbox violation if the enum
value inside the sandbox was modified by an attacker. Simply adding an
UNREACHABLE fixes this. See also issue 326086002 for more details about
this kind of pattern.

Bug: chromium:332475841
Change-Id: Ib276abe0954d2c76d1021f35eb25f69e2e77ebfe