[wasm][arm] Avoid DCHECK for statically OOB function indexes
Since https://crrev.com/c/5401777 we optimize the dispatch table access
for static indexes on indirect calls. For very large indexes which would
always trigger the bounds check (and trap) we then computed a dispatch
table offset above kMaxInt, which triggered a DCHECK on arm.
In release builds this would lead to the generation of wrong code, but
this code would never get executed because we always hit the bounds
check before.

This CL refactors the code a bit to detect static OOB accesses in even
more cases, and avoids the DCHECK by just using an offset of 0, which is
OK because the code will never be executed anyway.

Just not generating any code is difficult because we need to leave the
value stack in a consistent state, so we just keep generating the
unreachable code.

R=ahaas@chromium.org

Fixed: chromium:332572762
Change-Id: I310135ac188701e780d8420e3eee8e9c1c3f99ef