[wasm] Remove backlink from internal function to funcref
The WasmFuncRef is the in-sandbox wrapper around the trusted
WasmInternalFunction. They currently link each other, which makes it
difficult to allocate the two together. Also, we mostly need to get from
the function reference to the internal function.

Hence this CL removes the link from the WasmInternalFunction back to the
WasmFuncRef. In order to get the WasmFuncRef from a WasmFunctionData, we
now link the WasmFuncRef from there instead of the WasmInternalFunction
directly. This requires one more indirection for js-to-wasm calls.

R=jkummerow@chromium.org

Bug: v8:14564
Change-Id: Ifab716a11e6db0ccad54890d50c31db348b11a65