[metrics] Fix code cache related histograms
CacheBehaviour::kHitIsolateCacheWhenProduceCodeCache was never emitted
since ScriptCompileTimerScope::set_producing_code_cache() was never
called. These samples were counted in the kHitIsolateCacheWhenNoCache
bucket.

Producing the code cache is a separate API call and we can't know fo
sure whether the embedder will make the API call or not. Currently,
Chromium always sets the no_cache_reason to
kNoCacheBecauseDeferredProduceCodeCache when it's going to produce the
code cache, so this CL uses that as a signal.

Change-Id: I0dc4cbdc2aebcb63f082700b0a3145014446e37f