[wasm] Serialize enabled features
This adds the set of enabled features to the serialized data, such that
mismatches in the enabled features (e.g. due to origin trials) are
detected when deserializing. On a mismatch, the cached data is simply
ignored and the module is recompiled from scratch. The cache entry might
then be overwritten with the new version.

R=jkummerow@chromium.org

Fixed: v8:14711
Change-Id: I79d01b984a16a886f49196d08b5ff102360d4ef6