[liftoff] Add break conditions in SpillRegister
This prevents the loop index from overflowing by breaking the loop
when the index is 0.

Bug: The issue describes the flaky test in the NodeJS on Windows ARM64: https://github.com/nodejs/node/issues/51822

After investigation, I've discovered that the corresponding loop
doesn't break which throws an exception while reading the variable
cache_state_.stack_state[idx]. I've fixed this by adding two break
conditions using the variable idx. This is tested against the NodeJS
issue and it fixes the flaky test.

Change-Id: I16c338a84c737cf1287d9f5ce9a2a1b29d5db752