[liftoff] Add break conditions in SpillRegister
This prevents the loop index from overflowing by breaking the loop
when the index is 0.

Bug: After initiating the build process with Visual Studio version 17.9,
an issue arose causing a test to fail within the NodeJS framework on the
ARM64 system:
https://github.com/nodejs/node/issues/51822

After further investigation, I've discovered that the corresponding loop
doesn't break which throws an exception while reading the variable
cache_state_.stack_state[idx]. I've fixed this by adding two break
conditions using the variable idx. This is fixed and tested against the
NodeJS issue that detected this problem.

Change-Id: I16c338a84c737cf1287d9f5ce9a2a1b29d5db752