[wasm][liftoff] Fix check for statically OOB
A classic off-by-one for computing whether a static index is statically
known to be OOB (introduced in https://crrev.com/c/5425022).
In most cases we would get a signature mismatch when reading the
signature OOB from the dispatch table, but in rare cases we also get
crashes because the signature check passes and we load a garbage pointer
from the dispatch table.

R=mliedtke@chromium.org

Bug: chromium:333457544
Change-Id: I34a6bc5082d73d79c14ee5b9ed0141ea6d6b1d0b