[sandbox] Remove shortcut for evacuated segments during EPT sweeping
Originally, we had this shortcut to avoid iterating over segments that
we knew would be completely empty after evacuation and sweeping.
However, now that we have managed resources, we still need to process
entries that contain them in these segments. And we need to be careful
when doing so: as the segment hasn't been evacuated yet when we iterate
over it, not all entries are dead. The previous logic introduced in
https://crrev.com/c/5438117 got this wrong. Since we now anyway need to
iterate over the segment, the shortcut is much less useful, and so this
CL simply removes it, which also simplifies some other code.

Bug: chromium:333493383
Change-Id: I5a1453f19f2854f675ad3ab9c480cf0f9372bb0a