[profiler] Use gc_state to detect GC in GetStackSample
With the LOGGING VMState, we might overwrite the GC VMState (if e.g. we
are logging a move event). This is a problem in the profiler, which was
avoiding doing deep analysis of samples in GC, since the pointers it
would look at could be invalid during evacuation.

Use the heap's recorded gc_state instead, which is more precise than the
VMState and won't be clobbered by VMState changes. This lets us also be
specific that we only care about moving gc states, which tightens the
predicate a bit.

Bug: v8:14713
Change-Id: I756cc8f4d402effc12492b5a3c436bd2ba111a3e