[wasm][multi-memory] Add multi-memory support to wasm-module-builder.cc
We want to support multi-memory in the fuzzers which use the
wasm-module-builder.cc.

Before this change, the builder created one memory (32-bit) by default
and no more memories could be added. This CL changes the builder to
contain a WasmMemory vector. No memory is created by default after this
CL.

Bug: v8:14674
Change-Id: I99ec7310db2bacbfc0f3841759385321ebee7807