[wasm] Change original WebAssembly object in InstallConditionalFeatures
Origin trials can be enabled at any point in time, even after user code
has already been executed. This means that user code could have already
modified or replaced the WebAssembly object when V8 tries to add
additional functionality to the WebAssembly object.

So far the additional functionality was added to the user-visible
WebAssembly object at the time the origin trial gets enabled. This
caused problems, because the user-visible WebAssembly object could be an
object that is not extensible, or that needs access checks.

With this CL, additional funcitonality is added to the WebAssembly
object stored in the native context. This should limit possible problems
due to JS edge cases, at the expense of not enabling origin trials for
user code that replaced the WebAssembly object. However, we think this
is okay.

R=jkummerow@chromium.org

Change-Id: Ie87d47a5e063bbc44b2e692115df833ff8ac05d5